requirements.txt
opencv-python
pillow
numpy
streamlit
matplotlib
pandas
torch
tensorflow
